Updates Order so that users and admin can view previous orders as well as admin can update the status manually.

+++ db/migrate/20221126203532_create_orders.rb
@@ -5,9 +5,9 @@ class CreateOrders < ActiveRecord::Migration[7.0]
       t.date :order_processed
       t.date :order_shipped
       t.string :order_status
-      t.decimal :sub_total
-      t.decimal :order_total
-      t.decimal :total_payed
+      t.integer :sub_total
+      t.integer :order_total
+      t.integer :total_payed
 
       t.timestamps
     end
